add:func to create task

+++ models/user.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String
+from sqlalchemy import Column, BigInteger, String
 from sqlalchemy.ext.declarative import declarative_base
 
 Base = declarative_base()
@@ -7,5 +7,5 @@ Base = declarative_base()
 class UserToken(Base):
     __tablename__ = 'user_tokens'
 
-    user_id = Column(Integer, primary_key=True)
+    user_id = Column(BigInteger, primary_key=True)
     token = Column(String, nullable=True)
